+++ src/InnovateFuture.Api/Profiles/AutoMapperProfile.cs
@@ -3,6 +3,9 @@ using AutoMapper;
 using InnovateFuture.Api.Controllers.Organisations;
 using InnovateFuture.Api.Controllers.Profiles;
 using InnovateFuture.Api.Controllers.Users;
+using InnovateFuture.Application.Auth.Commands.ConfirmEmail;
+using InnovateFuture.Application.Auth.Commands.Login;
+using InnovateFuture.Application.Auth.Commands.Register;
 using InnovateFuture.Application.Common.Models;
 using InnovateFuture.Application.Profiles.Commands.UpdateProfile;
 using InnovateFuture.Application.Users.Commands.CreateUser;
@@ -12,10 +15,6 @@ using InnovateFuture.Application.Organisations.Commands.CreateOrganisation;
 using InnovateFuture.Application.Organisations.Commands.UpdateOrganisation;
 using InnovateFuture.Application.Organisations.Queries.GetOrganisations;
 using InnovateFuture.Application.Profiles.Queries.GetProfiles;
-using InnovateFuture.Application.Auth.ConfirmEmail;
-using InnovateFuture.Application.Auth.Login;
-using InnovateFuture.Application.Auth.Register;
-using InnovateFuture.Application.Services.Auth.Query;
 using InnovateFuture.Domain.Entities;
 using InnovateFuture.Domain.Enums;
 using APIQueryProfileFilters = InnovateFuture.Api.Controllers.Profiles.QueryProfileFilters;
@@ -37,7 +36,7 @@ public class AutoMapperProfile: AMProfile
         CreateMap<RegisterOrganisationAdminRequest, RegisterOrganisationAdminCommand>();
         CreateMap<LoginRequest, LoginCommand>();
 
-        CreateMap<GetMeDto,GetMeResponse>();
+        CreateMap<Profile,GetMeResponse>();
         
         CreateMap<CreateUserRequest, CreateUserCommand>();
         /*
