+++ src/InnovateFuture.Application/Auth/Commands/Login/LoginValidator.cs
@@ -1,6 +1,6 @@
 using FluentValidation;
 
-namespace InnovateFuture.Application.Auth.Login;
+namespace InnovateFuture.Application.Auth.Commands.Login;
 
 public class LoginValidator:  AbstractValidator<LoginCommand>
 {
