+++ src/InnovateFuture.Application/Auth/Commands/Register/InviteUserCommand.cs
@@ -1,0 +1,6 @@
+namespace InnovateFuture.Application.Auth.Commands.Register;
+
+public class InviteUserCommand
+{
+    
+}
