+++ src/InnovateFuture.Application/Auth/Queries/GetMe/GetMeQuery.cs
@@ -1,0 +1,14 @@
+using InnovateFuture.Domain.Entities;
+using MediatR;
+
+namespace InnovateFuture.Application.Auth.Queries.GetMe;
+
+public class GetMeQuery: IRequest<Profile>
+{
+    public Guid ProfileId { get; }
+
+    public GetMeQuery(Guid profileId)
+    {
+        ProfileId = profileId;
+    }
+}
