--- src/InnovateFuture.Application/Auth/Register/InviteUserCommand.cs
@@ -1,6 +1,0 @@
-namespace InnovateFuture.Application.Auth.Register;
-
-public class InviteUserCommand
-{
-    
-}
