--- src/InnovateFuture.Application/Services/Auth/Query/GetMeDto.cs
@@ -1,8 +1,0 @@
-using InnovateFuture.Domain.Entities;
-
-namespace InnovateFuture.Application.Services.Auth.Query;
-
-public class GetMeDto
-{
-    public Profile DefaultProfile { get; set; }
-}
