--- src/InnovateFuture.Application/Services/Auth/Query/GetMeQuery.cs
@@ -1,13 +1,0 @@
-using MediatR;
-
-namespace InnovateFuture.Application.Services.Auth.Query;
-
-public class GetMeQuery: IRequest<GetMeDto>
-{
-    public Guid ProfileId { get; }
-
-    public GetMeQuery(Guid profileId)
-    {
-        ProfileId = profileId;
-    }
-}
